renamed vehicle card for dashboard list
